refactor: extract FileLock from sidecar_metadata to file_lock module (#377)
