+++ Firmware/webui/backend/lib/active_state.py
@@ -14,7 +14,7 @@ def load_active_state() -> dict | None:
     Returns the parsed dict, or None if no active state.
     """
     from mothbox_paths import CONFIG_DIR
-    from webui.backend.lib.sidecar_metadata import FileLock
+    from webui.backend.lib.file_lock import FileLock
 
     state_file = CONFIG_DIR / "active_state.json"
     if not state_file.exists():
