+++ Firmware/webui/cli/refresh_schedule.py
@@ -51,7 +51,7 @@ def save_active_state(
     Returns True on success, False on failure.
     """
     from mothbox_paths import CONFIG_DIR
-    from webui.backend.lib.sidecar_metadata import FileLock
+    from webui.backend.lib.file_lock import FileLock
 
     state_file = CONFIG_DIR / "active_state.json"
     try:
